Create code for blinking an LED on an Arduino board
This code uses the digitalWrite() function to turn the LED on and off, and the delay() function to pause the program for a specified amount of time. In this example, the LED will blink on and off every second. You can modify the delay() function's value to change the blink rate.